Implement the JobUpdater
The JobUpdater is a new component which makes sure that
the information in the jobs already submitted to the batch
system are up to date. At the moment this only
concern the value of priority through the CondorPlugin.

Additionally in the case of priority, it also ensures that
the value is in synch with WMBS and the Local Queue.